structured functions into a model file
